+++ relm4-macros/src/widgets/gen/init_widgets.rs
@@ -1,0 +1,14 @@
+use proc_macro2::TokenStream as TokenStream2;
+use quote::quote;
+
+use super::Widget;
+
+impl Widget {
+    pub fn init_widgets_stream(&self, stream: &mut TokenStream2) {
+        let name = &self.name;
+        let func = self.func.func_token_stream();
+        stream.extend(quote! {
+            let #name = #func;
+        });
+    }
+}